+++ Sources/KeyboardKit/Feedback/KeyboardFeedbackHandler.swift
@@ -40,12 +40,6 @@ public protocol KeyboardFeedbackHandler {
      try out how a gesture will behave for a certain action.
      */
     func triggerFeedback(for gesture: KeyboardGesture, on action: KeyboardAction)
-    
-    /**
-     Trigger feedback for when space is pressed so long that
-     it enters cursor drag state.
-     */
-    func triggerFeedbackForLongPressOnSpaceDragGesture()
  
     /**
      Trigger feedback for when a `gesture` is performed on a